requirements.txt
asgiref==3.5.2
asttokens==2.2.1
atomicwrites==1.4.1
attrs==22.1.0
backcall==0.2.0
certifi==2022.12.7
chardet==3.0.4
colorama==0.4.6
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
Django==3.2
django-debug-toolbar==2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
executing==1.2.0
Faker==12.0.1
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
idna==2.8
importlib-metadata==6.0.0
iniconfig==1.1.1
ipykernel==6.21.1
ipython==8.9.0
jedi==0.18.2
jupyter_client==8.0.2
jupyter_core==5.2.0
matplotlib-inline==0.1.6
mixer==7.1.2
more-itertools==9.0.0
nest-asyncio==1.5.6
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.2
pluggy==0.13.1
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
py==1.11.0
Pygments==2.14.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytils==0.3
pytz==2022.6
pywin32==305
pyzmq==25.0.0
requests==2.22.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.4.3
stack-data==0.6.2
toml==0.10.2
tornado==6.2
traitlets==5.9.0
urllib3==1.25.11
wcwidth==0.2.6
zipp==3.12.1
